+++ src/std/core/io/mod.rs
@@ -17,28 +17,20 @@ impl StdCoreIo {
 
     pub fn load(&mut self) {
         self.load_print();
-        self.load_println();
+        self.load_eprint();
+        self.load_warn();
+        self.load_panic();
+        self.load_exit();
+        self.load_read_num();
+        self.load_read_str();
+        self.load_read_char();
+        self.load_read_bool();
     }
 
     /// print(text) - prints text to the terminal
     pub fn load_print(&mut self) {
         func(
             "print",
-            1,
-            &mut self.env,
-            Rc::new(Wrapper {
-                0: Box::new(|args: &[LiteralType]| {
-                    print!("{}", args[0].to_string());
-                    LiteralType::Void
-                }),
-            }),
-        );
-    }
-
-    /// println(text) - prints text to the terminal with a newline
-    pub fn load_println(&mut self) {
-        func(
-            "println",
             1,
             &mut self.env,
             Rc::new(Wrapper {
@@ -58,21 +50,6 @@ impl StdCoreIo {
             &mut self.env,
             Rc::new(Wrapper {
                 0: Box::new(|args: &[LiteralType]| {
-                    print!("{}", args[0].to_string().red());
-                    LiteralType::Void
-                }),
-            }),
-        );
-    }
-
-    /// eprintln(text) - prints error text to the terminal with a newline
-    pub fn load_eprintln(&mut self) {
-        func(
-            "eprintln",
-            1,
-            &mut self.env,
-            Rc::new(Wrapper {
-                0: Box::new(|args: &[LiteralType]| {
                     println!("{}", args[0].to_string().red());
                     LiteralType::Void
                 }),
@@ -83,20 +60,6 @@ impl StdCoreIo {
     pub fn load_warn(&mut self) {
         func(
             "warn",
-            1,
-            &mut self.env,
-            Rc::new(Wrapper {
-                0: Box::new(|args: &[LiteralType]| {
-                    print!("{}", args[0].to_string().yellow());
-                    LiteralType::Void
-                }),
-            }),
-        );
-    }
-
-    pub fn load_warnln(&mut self) {
-        func(
-            "warnln",
             1,
             &mut self.env,
             Rc::new(Wrapper {
@@ -135,6 +98,69 @@ impl StdCoreIo {
                         }
                     }
                     exit(0);
+                }),
+            }),
+        );
+    }
+
+    pub fn load_read_str(&mut self) {
+        func(
+            "readstr",
+            0,
+            &mut self.env,
+            Rc::new(Wrapper {
+                0: Box::new(|_args: &[LiteralType]| {
+                    let mut input = String::new();
+                    std::io::stdin().read_line(&mut input).unwrap();
+                    LiteralType::String(input)
+                }),
+            }),
+        );
+    }
+
+    pub fn load_read_num(&mut self) {
+        func(
+            "read_num",
+            0,
+            &mut self.env,
+            Rc::new(Wrapper {
+                0: Box::new(|_args: &[LiteralType]| {
+                    let mut input = String::new();
+                    std::io::stdin().read_line(&mut input).unwrap();
+                    let input = input.trim().parse::<f32>().unwrap();
+                    LiteralType::Number(input)
+                }),
+            }),
+        );
+    }
+
+    pub fn load_read_bool(&mut self) {
+        func(
+            "read_bool",
+            0,
+            &mut self.env,
+            Rc::new(Wrapper {
+                0: Box::new(|_args: &[LiteralType]| {
+                    let mut input = String::new();
+                    std::io::stdin().read_line(&mut input).unwrap();
+                    let input = input.trim().parse::<bool>().unwrap();
+                    LiteralType::Boolean(input)
+                }),
+            }),
+        );
+    }
+
+    pub fn load_read_char(&mut self) {
+        func(
+            "read_char",
+            0,
+            &mut self.env,
+            Rc::new(Wrapper {
+                0: Box::new(|_args: &[LiteralType]| {
+                    let mut input = String::new();
+                    std::io::stdin().read_line(&mut input).unwrap();
+                    let input = input.trim().chars().next().unwrap();
+                    LiteralType::Char(input)
                 }),
             }),
         );